# Next event estimation