--- a/src/renderer/glsl/chunks/rayTraceCore.glsl
+++ b/src/renderer/glsl/chunks/rayTraceCore.glsl
@@ -4,7 +4,7 @@ export default `
   #define THICK_GLASS 2
   #define SHADOW_CATCHER 3
 
-  #define SAMPLES_PER_MATERIAL 8
+  #define SAMPLES_PER_MATERIAL 6
 
   const float IOR = 1.5;
   const float INV_IOR = 1.0 / IOR;
@@ -71,12 +71,12 @@ export default `
   struct Path {
     Ray ray;
     vec3 li;
-    vec3 albedo;
     float alpha;
     vec3 beta;
     bool specularBounce;
     bool abort;
     float misWeight;
+    float alphaFromLi;
   };
 
   uniform Camera camera;