--- a/src/renderer/glsl/chunks/rayTraceCore.glsl
+++ b/src/renderer/glsl/chunks/rayTraceCore.glsl
@@ -76,7 +76,7 @@ export default `
     bool specularBounce;
     bool abort;
     float misWeight;
-    float alphaFromLi;
+    bool luminanceOnly;
   };
 
   uniform Camera camera;