--- a/src/renderer/glsl/chunks/sampleGlassSpecular.glsl
+++ b/src/renderer/glsl/chunks/sampleGlassSpecular.glsl
@@ -4,17 +4,18 @@ export default `
 
 void sampleGlassSpecular(SurfaceInteraction si, int bounce, inout Path path) {
   bool lastBounce = bounce == BOUNCES;
-
   vec3 viewDir = -path.ray.d;
   float cosTheta = dot(si.normal, viewDir);
+
+  MaterialSamples samples = getRandomMaterialSamples();
+
+  float reflectionOrRefraction = samples.s1.x;
 
   float F = si.materialType == THIN_GLASS ?
     fresnelSchlick(abs(cosTheta), R0) : // thin glass
     fresnelSchlickTIR(cosTheta, R0, IOR); // thick glass
 
   vec3 lightDir;
-
-  float reflectionOrRefraction = randomSample();
 
   if (reflectionOrRefraction < F) {
     lightDir = reflect(-viewDir, si.normal);
@@ -27,11 +28,8 @@ void sampleGlassSpecular(SurfaceInteraction si, int bounce, inout Path path) {
 
   initRay(path.ray, si.position + EPS * lightDir, lightDir);
 
-  // advance sample index by unused stratified samples
-  const int usedSamples = 1;
-  sampleIndex += SAMPLES_PER_MATERIAL - usedSamples;
-
   path.li += lastBounce ? path.beta * sampleBackgroundFromDirection(lightDir) : vec3(0.0);
+  path.specularBounce = true;
 }
 
 #endif