--- a/src/renderer/glsl/chunks/sampleMaterial.glsl
+++ b/src/renderer/glsl/chunks/sampleMaterial.glsl
@@ -5,43 +5,45 @@ export default `
 
 void sampleMaterial(SurfaceInteraction si, int bounce, inout Path path) {
   bool lastBounce = bounce == BOUNCES;
-
   mat3 basis = orthonormalBasis(si.normal);
   vec3 viewDir = -path.ray.d;
 
-  vec2 diffuseOrSpecular = randomSampleVec2();
-  vec2 lightRand = randomSampleVec2();
-  vec2 bounceRand = randomSampleVec2();
+  MaterialSamples samples = getRandomMaterialSamples();
+
+  vec2 diffuseOrSpecular = samples.s1;
+  vec2 lightDirSample = samples.s2;
+  vec2 bounceDirSample = samples.s3;
+
+  // Step 1: Add direct illumination of the light source (the hdr map)
+  // On every bounce but the last, importance sample the light source
+  // On the last bounce, multiple importance sample the brdf AND the light source, determined by random var
 
   vec3 lightDir;
-  float cosThetaL;
-  vec3 brdf;
   vec2 uv;
   float lightPdf;
-  float scatteringPdf;
-  float orientation;
-
-  // Add path contribution
-  float liContrib = 1.0;
-
   bool brdfSample = false;
+
   if (lastBounce && diffuseOrSpecular.x < 0.5) {
-    lightDir = diffuseOrSpecular.y < mix(0.5, 0.0, si.metalness) ?
-      lightDirDiffuse(si.faceNormal, viewDir, basis, lightRand) :
-      lightDirSpecular(si.faceNormal, viewDir, basis, si.roughness, lightRand);
+    // reuse this sample by multiplying by 2 to bring sample from [0, 0.5), to [0, 1)
+    lightDir = 2.0 * diffuseOrSpecular.x < mix(0.5, 0.0, si.metalness) ?
+      lightDirDiffuse(si.faceNormal, viewDir, basis, lightDirSample) :
+      lightDirSpecular(si.faceNormal, viewDir, basis, si.roughness, lightDirSample);
 
     uv = cartesianToEquirect(lightDir);
     lightPdf = envmapPdf(uv);
     brdfSample = true;
   } else {
-    lightDir = sampleEnvmap(lightRand, uv, lightPdf);
+    lightDir = sampleEnvmap(lightDirSample, uv, lightPdf);
   }
 
-  cosThetaL = dot(si.normal, lightDir);
+  float cosThetaL = dot(si.normal, lightDir);
 
-  orientation = dot(si.faceNormal, viewDir) * cosThetaL;
+  float occluded = 1.0;
+
+  float orientation = dot(si.faceNormal, viewDir) * cosThetaL;
   if (orientation < 0.0) {
-    liContrib = 0.0;
+    // light dir points towards surface. invalid dir.
+    occluded = 0.0;
   }
 
   float diffuseWeight = 1.0;
@@ -51,56 +53,54 @@ void sampleMaterial(SurfaceInteraction si, int bounce, inout Path path) {
     if (lastBounce) {
       diffuseWeight = 0.0;
     } else {
-      liContrib = 0.0;
+      occluded = 0.0;
     }
   }
 
   vec3 irr = textureLinear(envmap, uv).rgb;
 
-  brdf = materialBrdf(si, viewDir, lightDir, cosThetaL, diffuseWeight, scatteringPdf);
+  float scatteringPdf;
+  vec3 brdf = materialBrdf(si, viewDir, lightDir, cosThetaL, diffuseWeight, scatteringPdf);
 
   float weight;
   if (lastBounce) {
     weight = brdfSample ?
       2.0 * powerHeuristic(scatteringPdf, lightPdf) / scatteringPdf :
       2.0 * powerHeuristic(lightPdf, scatteringPdf) / lightPdf;
-
   } else {
     weight = powerHeuristic(lightPdf, scatteringPdf) / lightPdf;
   }
 
-  vec3 li = liContrib * brdf * irr * abs(cosThetaL) * weight;
+  path.li += path.beta * occluded * brdf * irr * abs(cosThetaL) * weight;;
 
-  path.li += path.beta * li;
-
-  // Get new path direction
+  // Step 2: Setup ray direction for next bounce by importance sampling the BRDF
 
   if (lastBounce) {
     return;
   }
 
-  path.specularBounce = false;
-
   lightDir = diffuseOrSpecular.y < mix(0.5, 0.0, si.metalness) ?
-    lightDirDiffuse(si.faceNormal, viewDir, basis, bounceRand) :
-    lightDirSpecular(si.faceNormal, viewDir, basis, si.roughness, bounceRand);
+    lightDirDiffuse(si.faceNormal, viewDir, basis, bounceDirSample) :
+    lightDirSpecular(si.faceNormal, viewDir, basis, si.roughness, bounceDirSample);
 
   cosThetaL = dot(si.normal, lightDir);
 
-  brdf = materialBrdf(si, viewDir, lightDir, cosThetaL, 1.0, scatteringPdf);
+  orientation = dot(si.faceNormal, viewDir) * cosThetaL;
+  path.abort = orientation < 0.0;
 
-  path.beta *= abs(cosThetaL) * brdf / scatteringPdf;
+  if (path.abort) {
+    return;
+  }
+
+  brdf = materialBrdf(si, viewDir, lightDir, cosThetaL, 1.0, scatteringPdf);
 
   uv = cartesianToEquirect(lightDir);
   lightPdf = envmapPdf(uv);
+
   path.misWeight = powerHeuristic(scatteringPdf, lightPdf);
 
+  path.beta *= abs(cosThetaL) * brdf / scatteringPdf;
+
   initRay(path.ray, si.position + EPS * lightDir, lightDir);
-
-  // If new ray direction is pointing into the surface,
-  // the light path is physically impossible and we terminate the path.
-  orientation = dot(si.faceNormal, viewDir) * cosThetaL;
-  path.abort = orientation < 0.0;
 }
-
 `;