--- a/src/renderer/glsl/chunks/sampleShadowCatcher.glsl
+++ b/src/renderer/glsl/chunks/sampleShadowCatcher.glsl
@@ -2,130 +2,109 @@ export default `
 
 #ifdef USE_SHADOW_CATCHER
 
-float importanceSampleLightShadowCatcher(SurfaceInteraction si, vec3 viewDir, vec2 random, inout float alpha) {
-  float li;
-
-  float lightPdf;
-  vec2 uv;
-  vec3 lightDir = sampleEnvmap(random, uv, lightPdf);
-
-  float cosThetaL = dot(si.normal, lightDir);
-
-  float orientation = dot(si.faceNormal, viewDir) * cosThetaL;
-  if (orientation < 0.0) {
-    return li;
-  }
-
-  float occluded = 1.0;
-
-  Ray ray;
-  initRay(ray, si.position + EPS * lightDir, lightDir);
-  if (intersectSceneShadow(ray)) {
-    occluded = 0.0;
-  }
-
-  float irr = dot(luminance, textureLinear(envmap, uv).rgb);
-
-  // lambertian BRDF
-  float brdf = INVPI;
-  float scatteringPdf = abs(cosThetaL) * INVPI;
-
-  float weight = powerHeuristic(lightPdf, scatteringPdf);
-
-  float lightEq = irr * brdf * abs(cosThetaL) * weight / lightPdf;
-
-  alpha += lightEq;
-  li += occluded * lightEq;
-
-  return li;
-}
-
-float importanceSampleMaterialShadowCatcher(SurfaceInteraction si, vec3 viewDir, vec3 lightDir, inout float alpha) {
-  float li;
-
-  float cosThetaL = dot(si.normal, lightDir);
-
-  float orientation = dot(si.faceNormal, viewDir) * cosThetaL;
-  if (orientation < 0.0) {
-    return li;
-  }
-
-  float occluded = 1.0;
-
-  Ray ray;
-  initRay(ray, si.position + EPS * lightDir, lightDir);
-  if (intersectSceneShadow(ray)) {
-    occluded = 0.0;
-  }
-
-  vec2 uv = cartesianToEquirect(lightDir);
-
-  float lightPdf = envmapPdf(uv);
-
-  float irr = dot(luminance, textureLinear(envmap, uv).rgb);
-
-  // lambertian BRDF
-  float brdf = INVPI;
-  float scatteringPdf = abs(cosThetaL) * INVPI;
-
-  float weight = powerHeuristic(scatteringPdf, lightPdf);
-
-  float lightEq = irr * brdf * abs(cosThetaL) * weight / scatteringPdf;
-
-  alpha += lightEq;
-  li += occluded * lightEq;
-
-  return li;
-}
-
 void sampleShadowCatcher(SurfaceInteraction si, int bounce, inout Path path) {
+  bool lastBounce = bounce == BOUNCES;
+
   mat3 basis = orthonormalBasis(si.normal);
   vec3 viewDir = -path.ray.d;
-  vec3 color = bounce > 1 && !path.specularBounce ? sampleEnvmapFromDirection(-viewDir) : sampleBackgroundFromDirection(-viewDir);
 
-  vec3 lightDir = lightDirDiffuse(si.faceNormal, viewDir, basis, randomSampleVec2());
+  vec2 diffuseOrSpecular = randomSampleVec2();
+  vec2 lightRand = randomSampleVec2();
+  vec2 bounceRand = randomSampleVec2();
 
-  float alphaBounce = 0.0;
+  vec3 lightDir;
+  float cosThetaL;
+  vec3 brdf;
+  vec2 uv;
+  float lightPdf;
+  float scatteringPdf;
+  float orientation;
 
-  vec3 li = path.beta * color * (
-      importanceSampleLightShadowCatcher(si, viewDir, randomSampleVec2(), alphaBounce) +
-      importanceSampleMaterialShadowCatcher(si, viewDir, lightDir, alphaBounce)
-    );
+  // Add path contribution
 
-  // alphaBounce contains the lighting of the shadow catcher *without* shadows
-  alphaBounce = alphaBounce == 0.0 ? 1.0 : alphaBounce;
+  vec3 color = sampleEnvmapFromDirection(-viewDir);
 
-  // in post processing step, we divide by alpha to obtain the percentage of light relative to shadow for the shadow catcher
-  path.alpha *= alphaBounce;
+  float liContrib = 1.0;
 
-  // we only want the alpha division to affect the shadow catcher
-  // factor in alpha to the previous light, so that dividing by alpha with the previous light cancels out this contribution
-  path.li *= alphaBounce;
+  bool brdfSample = false;
+  if (lastBounce && diffuseOrSpecular.x < 0.5) {
+    // lightDir = diffuseOrSpecular.y < mix(0.5, 0.0, si.metalness) ?
+    //   lightDirDiffuse(si.faceNormal, viewDir, basis, lightRand) :
+    //   lightDirSpecular(si.faceNormal, viewDir, basis, si.roughness, lightRand);
+    lightDir = lightDirDiffuse(si.faceNormal, viewDir, basis, lightRand);
+    uv = cartesianToEquirect(lightDir);
+    lightPdf = envmapPdf(uv);
+    brdfSample = true;
+  } else {
+    lightDir = sampleEnvmap(lightRand, uv, lightPdf);
+  }
 
-  // add path contribution
-  path.li += li;
+  cosThetaL = dot(si.normal, lightDir);
+
+  orientation = dot(si.faceNormal, viewDir) * cosThetaL;
+  if (orientation < 0.0) {
+    liContrib = 0.0;
+  }
+
+  float visible = 1.0;
+  initRay(path.ray, si.position + EPS * lightDir, lightDir);
+  if (intersectSceneShadow(path.ray)) {
+    visible = 0.0;
+  }
+
+  float irr = dot(luminance, textureLinear(envmap, uv).rgb);
+
+  // brdf = materialBrdf(si, viewDir, lightDir, cosThetaL, 1.0, scatteringPdf);
+  brdf = vec3(INVPI);
+  scatteringPdf = abs(cosThetaL) * INVPI;
+
+  float weight;
+  if (lastBounce) {
+    weight = brdfSample ?
+      2.0 * powerHeuristic(scatteringPdf, lightPdf) / scatteringPdf :
+      2.0 * powerHeuristic(lightPdf, scatteringPdf) / lightPdf;
+  } else {
+    weight = powerHeuristic(lightPdf, scatteringPdf) / lightPdf;
+  }
+
+  float liEq = liContrib * brdf.r * irr * abs(cosThetaL) * weight;
+  float li = visible * liEq;
+
+  path.li += path.beta * color * li;
+  path.alpha *= liEq;
 
   // Get new path direction
 
-  lightDir = lightDirDiffuse(si.faceNormal, viewDir, basis, randomSampleVec2());
+  if (lastBounce) {
+    return;
+  }
 
-  float cosThetaL = dot(si.normal, lightDir);
+  path.specularBounce = false;
+  path.alphaFromLi = 1.0;
 
-  // lambertian brdf with terms cancelled
-  path.beta *= color;
+  // lightDir = diffuseOrSpecular.y < mix(0.5, 0.0, si.metalness) ?
+  //   lightDirDiffuse(si.faceNormal, viewDir, basis, bounceRand) :
+  //   lightDirSpecular(si.faceNormal, viewDir, basis, si.roughness, bounceRand);
+  lightDir = lightDirDiffuse(si.faceNormal, viewDir, basis, bounceRand);
+
+  cosThetaL = dot(si.normal, lightDir);
+
+  // brdf = materialBrdf(si, viewDir, lightDir, cosThetaL, 1.0, scatteringPdf);
+  brdf = vec3(INVPI);
+  scatteringPdf = abs(cosThetaL) * INVPI;
+
+  path.beta *= color * abs(cosThetaL) * brdf / scatteringPdf;
+
+  uv = cartesianToEquirect(lightDir);
+  lightPdf = envmapPdf(uv);
+  path.misWeight = powerHeuristic(scatteringPdf, lightPdf);
 
   initRay(path.ray, si.position + EPS * lightDir, lightDir);
 
   // If new ray direction is pointing into the surface,
   // the light path is physically impossible and we terminate the path.
-  float orientation = dot(si.faceNormal, viewDir) * cosThetaL;
+  orientation = dot(si.faceNormal, viewDir) * cosThetaL;
   path.abort = orientation < 0.0;
-
-  path.specularBounce = false;
-
-  // advance dimension index by unused stratified samples
-  const int usedSamples = 6;
-  sampleIndex += SAMPLES_PER_MATERIAL - usedSamples;
 }
 
 #endif