… src/renderer/glsl/rayTrace.frag