## Brief Description
This PR restructures the path tracing algorithm. In particular, we reduce the number of ray intersections leading to a 2x performance speedup, and a noticeable increase in convergence. This also means the renderer is more responsive, and we're able to render preview buffers at a higher resolution. Most importantly, speeding up the ray tracing pass by 2x lets us spend more time per frame for the spatial denoising filter coming down the line.

**Details**
Our path tracing algorithm was based off the one described in [pbrt](http://www.pbr-book.org/3ed-2018/Light_Transport_I_Surface_Reflection/Path_Tracing.html).

For each bounce:
1. Get direct illumination by importance sampling the light source (the hdr map)
2. Get direct illumination by importance sampling the material brdf.
3. Combine both results using multiple importance sampling (with the power heuristic)
4. Importance sample the brdf to start the next bounce

(1), (2) require shadow ray intersections, and (4) requires a full scene ray intersection. If (4) ends up hitting a light source, we discard the result and terminate the path.

This results in a total of **3 intersections per bounce**

This PR restructures the algorithm to work as follows.

1. Get direct illumination by importance sampling the light source (the hdr map)
2. Importance sample the brdf to start the next bounce
3. If the interesction is
    3.a. *a light source*: Use the power heuristic to add this light's contribution and terminate the path.
    3.b. *a surface*: Start next bounce on this surface.

This results in a total of **2 intersections per bounce**. We're able to do this by effectively combining the (2) and (4) steps in the original algorithm. Combining the steps results in increased performance per frame, but with less information being calculated, resulting in slower convergence per frame. However, the increase in performance outweighs the loss in convergence per frame, leading to a net *increase in convergence* over real time.

This technique is similar to the "next event estimation" technique described in 
* http://www.cs.uu.nl/docs/vakken/magr/2015-2016/slides/lecture%2008%20-%20variance%20reduction.pdf
* https://www.youtube.com/watch?v=bGU_JVyvzLA

These sources don't explicitly mention combining steps (2) and (4), but [these notes](http://www.cs.cornell.edu/courses/cs6630/2012sp/notes/07pathtr-notes.pdf) mention the idea in the section, *Version 1.1*

**Note**
This PR introduces a regression to the shadow catcher. Specifically on non-metallic, specular surfaces, the reflection of the shadow catcher appears somewhat brighter than it should be, with some (but not all) hdr maps. This issue stems from the shadow catcher itself (and not from next event estimation directly), but I've been unable to find a fix for the bug. The issue is subtle enough that it shouldn't be noticeable in most cases.


## Image(s) or GIFs (if applicable)
Master: 10 seconds of rendering - 15 fps
![master 15 fps](https://user-images.githubusercontent.com/4411068/74390488-4ed21980-4db6-11ea-95a3-42e9c38ad0c1.png)
Next event estimation: 10 seconds of rendering - 30 fps
![nee 30 fps](https://user-images.githubusercontent.com/4411068/74390506-5db8cc00-4db6-11ea-9eef-0aa8ef11a8dc.png)

Both renders have stratified sampling *disabled* to increase similarity between each version's random sampling.

The next event estimation version runs at twice the framerate, and has noticeable less noise after running for 10 seconds.

## Pull Request Guidelines
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] I have added pull requests labels which describe my contribution.
- [x] All existing tests passed.
    - [ ] I have added tests to cover my changes, of which pass.
- [x] I have [compared](https://github.com/hoverinc/ray-tracing-renderer/wiki/Contributing#comparing-changes) the render output of my branch to `master`.
- [x] My code has passed the ESLint configuration for this project.
- [ ] My change requires modifications to the documentation.
    - [ ] I have updated the documentation accordingly.